Improve validation error messages
Improve de validation error messages for more descriptive errors message
to the frontend application.

Translate messages to portuguese so the front end
application does not need to handle them.

Return the user token for the frontend to be able to authenticate
the user and redirect him for the app home.